# docs/swagger.yaml
basePath: development
definitions:
  middleware-mmksi_dsf_mrp_response.Vehicle:
    properties:
      BrandId:
        type: integer
      BrandName:
        type: string
      Models:
        items:
          $ref: '#/definitions/response.VehicleModel'
        type: array
    type: object
  middleware-mmksi_mmksi_response.Vehicle:
    properties:
      ID:
        type: integer
      VehicleModel_S1:
        type: string
      VehicleType:
        type: string
    type: object
...
  request.PredictionRequest:
    properties:
      BRAND:
        type: string
      CITY:
        type: string
      COLOR:
        type: string
      COMPANY:
        type: string
      DISTANCE:
        type: integer
      MODEL:
        type: string
      PROVINCE:
        type: string
      TIPE_PENJUAL:
        type: string
      TRANSMISSION:
        type: string
      VARIANT:
        type: string
      YEAR:
        type: integer
    type: object
  request.RefreshTokenAWSRequest:
    properties:
      refresh_token:
        type: string
      username:
        type: string
    type: object
...
  request.TokenAWSRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  request.VehicleRequest:
    properties:
      find:
        items:
          $ref: '#/definitions/request.findVehicleRequest'
        type: array
      pages:
        type: integer
      sort:
        items:
          $ref: '#/definitions/request.sortVehicleRequest'
        type: array
    type: object
...
  request.findVehicleRequest:
    properties:
      MatchType:
        type: integer
      PropertyName:
        type: string
      PropertyValue:
        type: string
      SqlOperation:
        type: integer
    type: object
...
  request.sortVehicleRequest:
    properties:
      SortColumn:
        type: string
      SortDirection:
        type: integer
    type: object
...
  response.City:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error_id:
        type: integer
      message:
        additionalProperties:
          type: string
        type: object
    type: object
  response.GetRegionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.Province'
        type: array
    type: object
  response.GetVehiclesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/middleware-mmksi_dsf_mrp_response.Vehicle'
        type: array
    type: object
...
  response.PredictionResponse:
    properties:
      AveragePrice:
        type: number
      MaxPrice:
        type: integer
      MedianPrice:
        type: integer
      MinPrice:
        type: integer
      ProvinceId:
        type: integer
    type: object
  response.Province:
    properties:
      Cities:
        items:
          $ref: '#/definitions/response.City'
        type: array
      ProvinceId:
        type: integer
      ProvinceName:
        type: string
    type: object
...
  response.TokenAWSResponse:
    properties:
      id_token:
        type: string
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  response.Variant:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  response.VehicleColor:
    properties:
      ColorCode:
        type: string
      ColorDescription:
        type: string
      ID:
        type: integer
      VehicleModel_S1:
        type: string
      VehicleType:
        type: string
    type: object
  response.VehicleColorResponse:
    properties:
      lst:
        items:
          $ref: '#/definitions/response.VehicleColor'
        type: array
      messages:
        items:
          $ref: '#/definitions/response.VehicleMessageResponse'
        type: array
      total:
        type: integer
    type: object
  response.VehicleMessageResponse:
    properties:
      ErrorCode:
        type: integer
      ErrorMessage:
        type: string
    type: object
  response.VehicleModel:
    properties:
      ModelId:
        type: integer
      ModelName:
        type: string
      Variants:
        items:
          $ref: '#/definitions/response.Variant'
        type: array
    type: object
  response.VehicleResponse:
    properties:
      lst:
        items:
          $ref: '#/definitions/middleware-mmksi_mmksi_response.Vehicle'
        type: array
      messages:
        items:
          $ref: '#/definitions/response.VehicleMessageResponse'
        type: array
      total:
        type: integer
    type: object
host: xuu77ziiri.execute-api.us-east-2.amazonaws.com/
info:
  contact: {}
  description: MMKSI Middleware API Documentation
  title: MMKSI Middleware API Documentation
  version: "1.0"
paths:
  /auth/resignin:
    post:
      description: Authenticates a user and provides a JWT to Authorize API calls
      parameters:
      - description: Request Body
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/request.RefreshTokenAWSRequest'
      - description: Company
        in: header
        name: company
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TokenAWSResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Provides a Refresh Token
      tags:
      - Token
  /auth/signin:
    post:
      description: Authenticates a user and provides a JWT to Authorize API calls
      parameters:
      - description: Request Body
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/request.TokenAWSRequest'
      - description: Company
        in: header
        name: company
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TokenAWSResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Provides a Sign In
      tags:
      - Token
...
  /dsf/tradein/prediction:
    post:
      description: Get Prediction
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - description: Request Body
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/request.PredictionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PredictionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Prediction
      tags:
      - MRP Tradein
  /dsf/tradein/regions:
    get:
      description: Get Regions
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - in: query
        name: province
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetRegionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Regions
      tags:
      - MRP Tradein
  /dsf/tradein/vehicles:
    get:
      description: Get Vehicle
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - in: query
        name: brandName
        type: string
      - in: query
        name: modelName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetVehiclesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Vehicle
      tags:
      - MRP Tradein
  /mmksi/getData:
    post:
      description: Get Vehicle from MMKSI
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - description: Vehicle
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/request.VehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.VehicleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Vehicle
      tags:
      - MMKSI Vehicle
  /mmksi/vehicle:
    post:
      description: Get Vehicle Color from MMKSI
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - description: Vehicle Color
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/request.VehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.VehicleColorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Vehicle Color
      tags:
      - MMKSI Vehicle
schemes:
- https
- http
swagger: "2.0"